requirements.txt
certifi==2026.2.25
clickhouse-connect==0.15.1
clickhouse-driver==0.2.10
greenlet==3.3.2
lz4==4.4.5
numpy==2.4.4
pandas==3.0.2
psycopg2-binary==2.9.11
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
zstandard==0.25.0
